Begin Implementation of Custom AI
Creates custom model to diagnose user diseases
Creates code to create future models
Begin NOSdiag (Not Outside Service Diagnosis) to be more accurate and stop reliance on outside AI
Reorganizes AIBot Directory
