--- AIBot/NLPmodel.md
@@ -1,3 +1,0 @@
-# Here is the link to the Natural Language Processing AI Model that is being used
-
-https://colab.research.google.com/drive/1ZM6RTKgEtkeVuRn0yubh6TxCeMlTogaE?usp=sharing
