+++ AIBot/chatgpt_words.txt
@@ -1,0 +1,1 @@
+It sounds like you may be experiencing a respiratory issue along with some swelling in your legs. I recommend seeking medical attention as soon as possible to get a proper evaluation and treatment.
