+++ AIBot/documentation/GemmaResearch.md
@@ -1,0 +1,12 @@
+# Gemma
+
+### Helpful Links
+* https://ai.google.dev/gemma/docs/setup
+* https://ai.google.dev/gemma/docs/get_started
+* https://huggingface.co/google/gemma-2b
+
+### Description
+Gemma is a family of lightweight, state-of-the-art open models from Google, built from the same research and technology used to create the Gemini models. They are text-to-text, decoder-only large language models, available in English, with open weights, pre-trained variants, and instruction-tuned variants. Gemma models are well-suited for a variety of text generation tasks, including question answering, summarization, and reasoning. Their relatively small size makes it possible to deploy them in environments with limited resources such as a laptop, desktop or your own cloud infrastructure, democratizing access to state of the art AI models and helping foster innovation for everyone.
+
+### Usage
+Below we share some code snippets on how to get quickly started with running the model. First make sure to pip install -U transformers, then copy the snippet from the section that is relevant for your usecase.
