+++ AIBot/documentation/datasets.md
@@ -1,0 +1,28 @@
+# Datasets
+
+## Useable
+### Disease and Symptoms
+* https://www.kaggle.com/datasets/choongqianzheng/disease-and-symptoms-dataset
+* https://www.kaggle.com/datasets/uom190346a/disease-symptoms-and-patient-profile-dataset/data
+* https://www.kaggle.com/datasets/pjmathematician/diseases-and-symptoms/data
+* https://www.kaggle.com/datasets/kaushil268/disease-prediction-using-machine-learning
+* https://www.kaggle.com/datasets/abbotpatcher/respiratory-symptoms-and-treatment
+
+* https://huggingface.co/datasets/QuyenAnhDE/Diseases_Symptoms/viewer/default/train?p=1
+* https://huggingface.co/datasets/dhivyeshrk/Disease-Symptom-Extensive-Clean
+* https://www.kaggle.com/datasets/itachi9604/disease-symptom-description-dataset?select=dataset.csv
+
+
+
+
+## ON HOLD
+### Cardiovascular information
+* https://www.kaggle.com/datasets/mirzahasnine/heart-disease-dataset
+  
+### Boolean Cardiovascular
+* https://www.kaggle.com/datasets/sulianova/cardiovascular-disease-dataset
+* https://www.kaggle.com/datasets/jocelyndumlao/cardiovascular-disease-dataset
+* https://www.kaggle.com/datasets/akshatshaw7/cardiovascular-disease-dataset
+* https://www.kaggle.com/datasets/christofel04/cardiovascular-study-dataset-predict-heart-disea?select=train.csv
+* https://www.kaggle.com/datasets/aiaiaidavid/cardio-data-dv13032020
+
