--- AIBot/research.md
@@ -1,27 +1,0 @@
-# Research
-There seems to be 4 options for how we may create the chat bot
-- Online API Service
-- Custom Large Language Model (LLM)
-- Adapted Large Language Model (LLM)
-- Non-AI 
-
-They each have different upsides and downsides
-
-### API Service
-OpenAI offers a variety of models for use. They are able to be customized, but use of them requires payment. Should not be too expensive, but its a barrier none the less.
-
-This also means the program will require constant internet access.
-
-### Custom LLM
-Needs an extremely large data set and training time. Very complicated to set up and very long training time. Seems the most in line with the spirit of the project but does truly seem out of our scope.
-
-Very indepth tutorial that shows the complexity. [link](https://www.youtube.com/watch?v=UU1WVnMk4E8)
-
-### Adapted LLM 
-More doable than a fully custom one and can be hosted locally. Many options to choose from but storing user data may be hard.
-
-### Non-AI Chat Bot
-More traditional bot we can cater responses for. Easy to set up and customizable, but will be extremely hard to expand for more diagnosises. Wouldn't do any actual thinking.
-Storing data will be easy but responses are generic
-
-This video shows a general tutorial how to set up one. [link](https://www.youtube.com/watch?v=CkkjXTER2KE)
