--- AIBot/responses.md
@@ -1,23 +1,0 @@
-
-Response: The most likely diagnosis is _____due to _____which can be temporarily treated with__
-
-Response: The likely diagnosis is ____ with ___ characterized by____
-Response: _____ should be considered, especially given the patient's history _____
-
-Response:  you have a chance of ___which can be temporarily cured with ____
-
-Responses: Suspected diagnosis is___, Please list more symptoms to narrow down the exact diagnosis. 
-
-Responses: ___can be symptoms of potentially serious conditions, such as __While I cannot provide a definitive diagnosis these symptoms should not be taken lightly. I strongly recommend seeking immediate medical attention, preferably at an emergency department, for a comprehensive evaluation. Timely diagnosis and treatment can be crucial in preventing further complications or permanent damage.
-
-Responses: Based on the limited information provided, ___could potentially be caused by various conditions, such as ___ , Do you have any other symptoms like___
-
-Responses: ____ can have various potential causes. To better understand what might be causing_____, I would need to ask you some additional questions, such as:
-
-When did the ___ first appear, and how quickly has it spread?
-
-Does the ____ have any specific patterns or characteristics (list physical symptoms, etc.)?
-
-Have you recently started any new medications, used new products, or been exposed to anything that could have caused an allergic reaction?
-
-Do you have any other symptoms like____
